Extracted firestore db service
