+++ firebase/build.gradle
@@ -1,0 +1,3 @@
+task assemble(){}
+task clean(){}
+task check(){}
