+++ firebase/firestore-db-service/README.md
@@ -1,0 +1,5 @@
+# Meant to stay only temporarily
+
+This module should disappear after having refactored this module properly.
+At the moment this module loads schedule related, twitter related, speaker related, venue related data, etc. from firestore.
+This functionality should be split into smaller classes and been moved into the corresponding service or repository module.
