+++ firebase/firestore-db-service/build.gradle
@@ -1,0 +1,57 @@
+buildscript {
+    repositories {
+        google()
+        jcenter()
+    }
+
+    apply from: '../../dependencies.gradle'
+
+    dependencies {
+        classpath buildScript.gradleStaticAnalysisPlugin
+        classpath gradlePlugins.kotlin
+    }
+}
+
+apply plugin: 'com.android.library'
+apply plugin: 'kotlin-android'
+
+android {
+    compileSdkVersion Integer.parseInt(project.COMPILE_SDK_VERSION)
+
+    defaultConfig {
+        minSdkVersion Integer.parseInt(project.MIN_SDK_VERSION)
+        targetSdkVersion Integer.parseInt(project.TARGET_SDK_VERSION)
+        versionCode Integer.parseInt(project.VERSION_CODE)
+        versionName project.VERSION_NAME
+
+        testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
+
+    }
+
+    buildTypes {
+        release {
+            minifyEnabled false
+            proguardFiles getDefaultProguardFile('proguard-android.txt'), 'proguard-rules.pro'
+        }
+    }
+
+}
+
+dependencies {
+    implementation fileTree(dir: 'libs', include: ['*.jar'])
+    api libraries.app.rxJava
+    api libraries.app.joda
+
+    implementation(libraries.app.firebase) {
+        exclude group: 'com.android.support'
+    }
+    implementation libraries.app.firebaseAuth
+    implementation libraries.app.firebaseConfig
+    implementation libraries.app.firebaseDb
+    implementation libraries.app.firebaseFirestore
+    implementation libraries.app.kotlin
+
+    testImplementation 'junit:junit:4.12'
+    androidTestImplementation 'com.android.support.test:runner:1.0.2'
+    androidTestImplementation 'com.android.support.test.espresso:espresso-core:3.0.2'
+}
