+++ firebase/firestore-db-service/src/main/java/net/squanchy/service/firebase/model/conferenceinfo/FirestoreConferenceInfo.kt
