+++ firebase/firestore-db-service/src/test/java/net/squanchy/service/firebase/ExampleUnitTest.java
@@ -1,0 +1,18 @@
+package net.squanchy.service.firebase;
+
+import org.junit.Test;
+
+import static org.junit.Assert.*;
+
+/**
+ * Example local unit test, which will execute on the development machine (host).
+ *
+ * @see <a href="http://d.android.com/tools/testing">Testing documentation</a>
+ */
+public class ExampleUnitTest {
+
+    @Test
+    public void addition_isCorrect() {
+        assertEquals(4, 2 + 2);
+    }
+}
