--- schedule/schedule-domain/src/main/java/com/squanchy/schedule/domain/view/Day.kt
@@ -1,5 +1,0 @@
-package com.squanchy.schedule.domain.view
-
-import org.joda.time.LocalDate
-
-data class Day(val id: String, val date: LocalDate)
