--- schedule/schedule-domain/src/main/java/com/squanchy/schedule/domain/view/SchedulePage.kt
@@ -1,9 +1,0 @@
-package com.squanchy.schedule.domain.view
-
-import org.joda.time.LocalDate
-
-data class SchedulePage(
-    val dayId: String,
-    val date: LocalDate,
-    val events: List<Event>
-)
