+++ schedule/schedule-domain/src/main/java/net/squanchy/schedule/domain/view/Day.kt
@@ -1,0 +1,5 @@
+package net.squanchy.schedule.domain.view
+
+import org.joda.time.LocalDate
+
+data class Day(val id: String, val date: LocalDate)
