+++ schedule/schedule-domain/src/main/java/net/squanchy/schedule/domain/view/Place.kt
@@ -1,0 +1,10 @@
+package net.squanchy.schedule.domain.view
+
+import arrow.core.Option
+
+data class Place(
+    val id: String,
+    val name: String,
+    val floor: Option<String>,
+    val position: Int
+)
