+++ schedule/schedule-domain/src/main/java/net/squanchy/schedule/domain/view/SchedulePage.kt
@@ -1,0 +1,9 @@
+package net.squanchy.schedule.domain.view
+
+import org.joda.time.LocalDate
+
+data class SchedulePage(
+    val dayId: String,
+    val date: LocalDate,
+    val events: List<Event>
+)
