+++ schedule/schedule-domain/src/main/java/net/squanchy/schedule/domain/view/Track.kt
@@ -1,0 +1,10 @@
+package net.squanchy.schedule.domain.view
+
+data class Track(
+    val id: String,
+    val numericId: Long,
+    val name: String,
+    val accentColor: String? = null,
+    val textColor: String? = null,
+    val iconUrl: String? = null
+)
