+++ schedule/schedule-service-firestore/.gitignore
@@ -1,0 +1,1 @@
+/build
