+++ schedule/schedule-service-firestore/src/main/AndroidManifest.xml
@@ -1,0 +1,2 @@
+<manifest xmlns:android="http://schemas.android.com/apk/res/android"
+  package="com.hannesdorfmann.schedule_service_firestore" />
