+++ schedule/schedule-service/src/main/java/net/squanchy/schedule/ScheduleService.kt
@@ -1,6 +1,7 @@
 package net.squanchy.schedule
 
 import io.reactivex.Observable
+import net.squanchy.schedule.domain.view.Schedule
 
 
 interface ScheduleService {
