+++ team-props/android-code-quality.gradle
@@ -1,20 +1,13 @@
-def justFolders = ["schedule"]
 
-task clean(type: Delete, dependsOn: subprojects.findAll { p ->
-    !justFolders.contains(p.name)
-}.collect { p -> "$p.name:clean" }) {
+task clean(type: Delete, dependsOn: subprojects.collect { p -> "$p.name:clean" }) {
     delete project.buildDir
 }
 
-task check(dependsOn: subprojects.findAll { p ->
-    !justFolders.contains(p.name)
-}.collect { p ->
+task check(dependsOn: subprojects.collect { p ->
     "$p.name:check"
 })
 
-task assemble(dependsOn: subprojects.findAll { p ->
-    !justFolders.contains(p.name)
-}.collect { p ->
+task assemble(dependsOn: subprojects.collect { p ->
     "$p.name:assemble"
 })
